Overhauling how the FE is organized
