+++ backend/profiles/views.py
@@ -9,6 +9,7 @@ class UserListCreate(ViewSetMixin, generics.ListCreateAPIView):
     """
     API endpoint that allows users to be viewed or edited.
     """
+
     queryset = User.objects.all()
     permission_classes = [permissions.AllowAny]
     serializer_class = UserSerializer
@@ -18,6 +19,7 @@ class GroupListCreate(ViewSetMixin, generics.ListCreateAPIView):
     """
     API endpoint that allows groups to be viewed or edited.
     """
+
     queryset = Group.objects.all()
     permission_classes = [permissions.AllowAny]
     serializer_class = GroupSerializer
