+++ frontend/src/components/common/footer/index.css
