+++ frontend/src/components/common/footer/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFacebookSquare, faInstagram } from '@fortawesome/free-brands-svg-icons';
-import './footer.css';
+import './index.css';
 
 
 export default class footer extends Component {
