+++ frontend/src/components/common/home/index.css
