+++ frontend/src/components/common/home/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import CarouselComponent from './carousel'
 
-import "./home.css";
+import "./index.css";
 
 class Home extends Component {
   render() {
