+++ frontend/src/components/static_pages/team.jsx
